Lazy instantiation of monkeycube mobs

+++ code/modules/reagents/reagent_containers/food/mobcube.dm
@@ -4,14 +4,14 @@
 	icon_state = "monkeycube"
 	bitesize = 12
 	food_flags = FOOD_MEAT
-	var/mob/living/carbon/contained_mob = new /mob/living/carbon/monkey(null) //Storing it in nullspace is fine
+	var/mob/living/carbon/contained_mob = /mob/living/carbon/monkey // Either a path or an instance of mob/living/carbon
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/New()
 	..()
 	reagents.add_reagent(NUTRIMENT,10)
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/Destroy()
-	if(contained_mob)
+	if(contained_mob && !ispath(contained_mob))
 		qdel(contained_mob)
 		contained_mob = null
 	..()
@@ -35,11 +35,10 @@
 	if (istype(M, /mob/living/carbon/human))
 		//Do not try to understand.
 		var/obj/item/weapon/surprise = new/obj/item/weapon(M)
-		surprise.icon = contained_mob.icon
-		surprise.icon_state = contained_mob.icon_state
+		surprise.icon = initial(contained_mob.icon)
+		surprise.icon_state = initial(contained_mob.icon_state)
 		surprise.name = "malformed [contained_mob.name]"
 		surprise.desc = "Looks like \a very deformed [contained_mob.name], a little small for its kind. It shows no signs of life."
-		qdel(contained_mob)
 		surprise.transform *= 0.6
 		surprise.add_blood(M)
 		var/mob/living/carbon/human/H = M
@@ -56,19 +55,26 @@
 			E.embed(surprise)
 	else if (ismonkey(M))
 		M.visible_message("<span class='danger'>[M] suddenly tears in half!</span>")
-		contained_mob.forceMove(get_turf(M))
-		contained_mob.name = "malformed [initial(name)]"
-		contained_mob.transform *= 0.6
-		contained_mob.add_blood(M)
+		var/mob/living/carbon/target_mob
+		if(ispath(contained_mob))
+			target_mob = new contained_mob(get_turf(M))
+		else
+			target_mob = contained_mob
+			target_mob.forceMove(get_turf(M))
+		target_mob.name = "malformed [initial(name)]"
+		target_mob.transform *= 0.6
+		target_mob.add_blood(M)
 		M.gib()
 	..()
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/proc/Expand()
-
 	for(var/mob/M in viewers(src,7))
 		to_chat(M, "<span class='warning'>\The [src] expands!</span>")
-	contained_mob.forceMove(get_turf(src))
-	contained_mob = null
+	if(ispath(contained_mob))
+		new contained_mob(get_turf(src))
+	else
+		contained_mob.forceMove(get_turf(src))
+		contained_mob = null
 	qdel(src)
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/proc/Unwrap(mob/user as mob)
@@ -87,27 +93,27 @@
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/farwacube
 	name = "farwa cube"
-	contained_mob = new /mob/living/carbon/monkey/tajara(null)
+	contained_mob = /mob/living/carbon/monkey/tajara
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/wrapped/farwacube
 	name = "farwa cube"
-	contained_mob = new /mob/living/carbon/monkey/tajara(null)
+	contained_mob = /mob/living/carbon/monkey/tajara
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/stokcube
 	name = "stok cube"
-	contained_mob = new /mob/living/carbon/monkey/unathi(null)
+	contained_mob = /mob/living/carbon/monkey/unathi
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/wrapped/stokcube
 	name = "stok cube"
-	contained_mob = new /mob/living/carbon/monkey/unathi(null)
+	contained_mob = /mob/living/carbon/monkey/unathi
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/neaeracube
 	name = "neaera cube"
-	contained_mob = new /mob/living/carbon/monkey/skrell(null)
+	contained_mob = /mob/living/carbon/monkey/skrell
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/wrapped/neaeracube
 	name = "neaera cube"
-	contained_mob = new /mob/living/carbon/monkey/skrell(null)
+	contained_mob = /mob/living/carbon/monkey/skrell
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/humancube
 	name = "humanoid cube"
